create simple abstract implementation for decorators and use it across update and query
